escape __VIRTUAL_ENV__ path.
Fixes issue https://github.com/pypa/virtualenv/issues/53 in bash and shell by using the printf %q "shell quote" formatter, which escapes as appropriate for a given environment. Falls back to quoting some characters that universally need to be escaped. Fallback is needed because there are several printf implementations, not all of which are guaranteed to have the %q formatter (though bash always should).